requirements.txt
pymilvus==2.5.6
python-dotenv==1.1.0
numpy==1.26.4
fastapi==0.110.0
uvicorn==0.27.1